Fix terraform plan file should not be deleted
issue: https://linear.app/cloudposse/issue/DEV-2825/atmos-terraform-show-deletes-planfile

+++ internal/exec/terraform.go
@@ -478,7 +478,7 @@ func ExecuteTerraform(info schema.ConfigAndStacksInfo) error {
 	}
 
 	// Clean up
-	if info.SubCommand != "plan" && info.PlanFile == "" {
+	if info.SubCommand != "plan" && info.SubCommand != "show" && info.PlanFile == "" {
 		planFilePath := constructTerraformComponentPlanfilePath(cliConfig, info)
 		_ = os.Remove(planFilePath)
 	}
